Move analytics classes into Analytics namespace
